+++ pga3dphysics/jvm/src/test/scala/me/kright/gametools/pga3d/physics/Pga3dInertiaGenerators.scala
@@ -57,7 +57,6 @@ object Pga3dInertiaGenerators:
     ).flatMap { inertia =>
       Gen.oneOf(
         inertia,
-        inertia.toPrecomputed,
         inertia.toInertiaMovedLocal,
         inertia.toFastestRepresentation,
         inertia.toSummable,