+++ pga3dphysics/jvm/src/test/scala/me/kright/gametools/pga3d/physics/Pga3dInertiaSimpleTest.scala
@@ -92,7 +92,6 @@ class Pga3dInertiaSimpleTest extends AnyFunSuiteLike with ScalaCheckPropertyChec
       val representations = ArraySeq[Pga3dInertia](
         inertia.toInertiaLocal,
         inertia.toInertiaMovedLocal,
-        inertia.toPrecomputed,
         inertia.toMovedSimple,
         inertia.toFastestRepresentation
       )